ref: unified entities graph
